Address MSVC warnings.
Drive-by: comb ProcessFlag in cjxl

+++ lib/jxl/dec_frame.h
@@ -181,7 +181,7 @@ class FrameDecoder {
 
   size_t NextNumPassesToPause() const {
     auto it = std::upper_bound(passes_to_pause_.begin(), passes_to_pause_.end(),
-                               NumCompletePasses());
+                               static_cast<int>(NumCompletePasses()));
     return (it != passes_to_pause_.end() ? *it
                                          : std::numeric_limits<size_t>::max());
   }
